refactor(scheduling): delete dead router, port reminder hooks to gateway
services/scheduling/src/router.ts defined a tRPC router that nothing
imported. The gateway re-implements every appointment procedure
directly in services/api-gateway/src/routers/scheduling.ts, so the
scheduling-package router was unreachable from production traffic.
Two consequences:

1. ~500 LOC of duplicated procedure definitions kept (badly) in sync
   with the gateway version that actually ran.
2. The dead router was the only place that wired up issue #333's
   appointment reminders (scheduleReminders / cancelReminders). The
   gateway's appointment routes never called them, so reminder
   delivery has been silently broken for the entire production
   lifetime of the gateway scheduling endpoints — independently of
   the worker-not-started bug fixed in #984 / PR #990.

This commit closes both gaps:

- Ports scheduleReminders / cancelReminders hooks into the gateway's
  scheduling.ts create / cancel / reschedule procedures, preserving
  the dead router's contract (best-effort, non-blocking, persist
  returned job IDs in a follow-up UPDATE, null them on cancel).
- Deletes services/scheduling/src/router.ts and the test file that
  covered the deleted router.
- Removes schedulingRouter / SchedulingRouter exports from
  services/scheduling/src/index.ts. The reminder helpers, the worker
  factory, and the worker server (started by services/scheduling/
  src/server.ts from #984) stay — those are still consumed.
- Updates caregiver-scope-matrix.test.ts to add
  DEFAULT_RETENTION_AGE_SECONDS to its redis-config mock, since the
  gateway's scheduling.ts now transitively imports reminders.ts.

Closes #1006
